+++ context/app/static/js/components/workspaces/hooks.js
@@ -1,0 +1,38 @@
+import { useContext, useState, useEffect } from 'react';
+
+import { AppContext } from 'js/components/Providers';
+
+function useWorkspacesList() {
+  const [workspacesList, setWorkspacesList] = useState([]);
+  // TODO: isLoading:
+  // const [isLoading, setIsLoading] = useState(true);
+
+  const { workspacesEndpoint, workspacesToken } = useContext(AppContext);
+
+  useEffect(() => {
+    async function getAndSetWorkspacesList() {
+      const response = await fetch(`${workspacesEndpoint}/workspaces`, {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+          'UWS-Authorization': `Token ${workspacesToken}`,
+        },
+      });
+
+      if (!response.ok) {
+        console.error('Workspaces API failed', response);
+        return;
+      }
+      const results = await response.json();
+
+      setWorkspacesList(results);
+      // TODO:
+      // setIsLoading(false);
+    }
+    getAndSetWorkspacesList();
+  }, [workspacesEndpoint, workspacesToken]);
+
+  return { workspacesList };
+}
+
+export { useWorkspacesList };
